Create #day021_Break_&_Continue.md
Day 21 : This file contains explanations and examples of the `break` and `continue` statements in Java. It covers their syntax and demonstrates how they can be used to control the flow of loops. The `break` statement is used to exit a loop or switch statement prematurely, while the `continue` statement is used to skip the current iteration of a loop and proceed with the next iteration.